Fixed problem with queries on smartphone table

+++ springSmartphoneShop/src/main/java/com/smartphoneShop/backend/service/UserService.java
@@ -8,11 +8,11 @@ public interface UserService {
 
     List<Users> getAllUsers();
 
-    void deleteById(long id);
+    void deleteById(int id);
 
     void updateUser(Users user);
 
-    Users findById(long id) throws Exception;
+    Users findById(int id) throws Exception;
 
     Users addUser(Users user);
 }
